10/11 21時時点進捗
