+++ Assets/Prefab/Enemy10(Bonus).prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 6514295398814519240}
   - component: {fileID: 4067544013864852461}
   - component: {fileID: 4017512423563098172}
+  - component: {fileID: 2189692323219629933}
   m_Layer: 11
   m_Name: Enemy10(Bonus)
   m_TagString: Enemy
@@ -172,3 +173,17 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 30
+  _power: 1
+--- !u!114 &2189692323219629933
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8781125943964791314}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8e5552d68ec2ac14995e0dece046a245, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHp: 100
