+++ Assets/Prefab/Enemy4(Shooter).prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 2402456300388049327}
   - component: {fileID: 5422587806203618926}
   - component: {fileID: 2883447396802500247}
+  - component: {fileID: 1907081687483839783}
   m_Layer: 10
   m_Name: Enemy4(Shooter)
   m_TagString: Enemy
@@ -172,3 +173,19 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 10
+  Player: {fileID: 2761608554920923673, guid: 2b6b41ae866e28e4b912f8a6ce5cf627, type: 3}
+  time: 0
+  _power: 1
+--- !u!114 &1907081687483839783
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7218224524036820350}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8e5552d68ec2ac14995e0dece046a245, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHp: 100
