+++ Assets/Prefab/Enemy6(LeftSnake).prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 769085427897163475}
   - component: {fileID: 9024537709586429898}
   - component: {fileID: 8248517789722741856}
+  - component: {fileID: 4102013686889445854}
   m_Layer: 10
   m_Name: Enemy6(LeftSnake)
   m_TagString: Enemy
@@ -174,3 +175,17 @@ MonoBehaviour:
   horizontalSpeed: 8
   verticalAmplitude: 10
   verticalFrequency: 5
+  _power: 1
+--- !u!114 &4102013686889445854
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 721788676681865757}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8e5552d68ec2ac14995e0dece046a245, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHp: 100
