+++ Assets/Prefab/Enemy8(RightSnake-halftempo).prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 772622924859519294}
   - component: {fileID: 9178674156261010389}
   - component: {fileID: 9027552273700391584}
+  - component: {fileID: 1022508557273857743}
   m_Layer: 10
   m_Name: Enemy8(RightSnake-halftempo)
   m_TagString: Enemy
@@ -175,3 +176,17 @@ MonoBehaviour:
   verticalAmplitude: 10
   verticalFrequency: 5
   phaseShift: 1.570796
+  _power: 1
+--- !u!114 &1022508557273857743
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 203159238664605855}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8e5552d68ec2ac14995e0dece046a245, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHp: 100
