+++ Assets/Prefab/Enemy9(LeftSnake-halftempo).prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 585922414013657349}
   - component: {fileID: 8593650666018975934}
   - component: {fileID: 7139289720644857426}
+  - component: {fileID: 2659356798920437144}
   m_Layer: 10
   m_Name: Enemy9(LeftSnake-halftempo)
   m_TagString: Enemy
@@ -175,3 +176,17 @@ MonoBehaviour:
   verticalAmplitude: 10
   verticalFrequency: 5
   phaseShift: 1.570796
+  _power: 1
+--- !u!114 &2659356798920437144
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1099553750580496013}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8e5552d68ec2ac14995e0dece046a245, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxHp: 100
