+++ Assets/Script/Enemy/Enemy2.cs
@@ -6,6 +6,7 @@ public class Enemy2: MonoBehaviour
 {
     Transform playerTr; // プレイヤーのTransform
     [SerializeField] float speed = 4; // 敵の動くスピード
+    [SerializeField] int _power;
 
     private void Start()
     {
@@ -27,6 +28,14 @@ public class Enemy2: MonoBehaviour
             transform.position,
             new Vector2(playerTr.position.x, playerTr.position.y),
             speed * Time.deltaTime);
+    }
+
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if (collision.gameObject.tag == "Player")
+        {
+            collision.gameObject.GetComponent<PlayerHP>().TakeDamage(_power);
+        }
     }
 
 }
