+++ Assets/Script/Enemy/Enemy6.cs.meta
