+++ Assets/Script/Player.cs
@@ -131,4 +131,5 @@ public class Player : MonoBehaviour
         _bRecastShot=false;
 
     }
+
 }
