+++ Assets/Script/PlayerHP.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fbb1a2df77c5aff4dad32c885a51fabd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
