+++ Assets/Script/Sword.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 84621a8fd62dc224aadecd0cfe8f8a48
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
